Builds are mostly correct now.
There's a couple of rought pieces I'd like to improve. Blog posts are
always loaded into memory and now I am wondering if maybe we should load
and build everything onto memory when it's first loaded. I want to
change the watch mode so it eventually only builds changed files and having
everything on memory for that could make it easier.