Improved file handling
This upgrades how we handle file changes and how these are used for hot reloading.

- We now debounce all file changes coming from the watcher, we do this on a per file
basis which should stop the server from recompiling a bunch when the file receives
multiple events.

- We now handle files created correctly. New files are added to the file tree and
compiled correctly without having to restart the server.

- Finally we switched from our barebones reloading mechanism to bundling livereload.
Live reload is generally a nicer experience, it handles css and img updates without
a full reload.